requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
WTForms>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
paramiko>=3.4.0
email-validator>=2.1.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
Flask-Talisman>=1.0.0
Flask-Limiter>=2.0.0
limits>=2.9.0
cryptography>=40.0
flask-sock>=0.7.0
gevent>=22.10.2
gevent-websocket>=0.10.1
websocket-client>=1.6.0
pycryptodome>=3.19.0
gunicorn>=21.2.0
